requirements.txt
ultralytics
pyserial
